backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9.9

